feat: remove commit command and add smart auto-staging to generate (#8)
