+++ src/cli.ts
@@ -5,7 +5,6 @@ import { load } from '@std/dotenv';
 import { red } from '@std/fmt/colors';
 
 import { handleGenerate } from './cmd/generate.ts';
-import { handleCommit } from './cmd/commit.ts';
 import { handleModel } from './cmd/model.ts';
 import { handleStatus } from './cmd/status.ts';
 import { handleVersion } from './cmd/version.ts';
@@ -86,16 +85,6 @@ cli
   .command('version', 'Show version information')
   .alias('v')
   .action(() => handleVersion());
-
-// Add commit command
-cli
-  .command('commit', 'Generate and commit changes with AI')
-  .alias('c')
-  .option('--model <model:string>', 'AI model to use')
-  .option('-p, --provider <provider:string>', 'AI provider to use')
-  .option('--staged', 'Only commit staged changes (default: stage all)')
-  .option('-d, --debug', 'Enable debug output')
-  .action((opts) => handleCommit(opts));
 
 // Add model command
 cli
